+++ rank3input_layer.py
@@ -3,7 +3,7 @@ import numpy as np
 class Rank3InputLayer():
     def __init__(self, shape):
         self.shape = shape
-        self.output = np.zeros(shape)
+        self.output = np.zeros(shape, dtype=np.float32)
 
     def set_input(self, ipt):
         if ipt.shape != self.shape:
